Moved utility functions from graphService to another file
